Move HashTable and Table initializer from file to the header so they can be used by other kernels. Change: 152999226
